LPS-91239 also run ddm form paragraph field upgrade for 7.0
